refactor: `domain` 패키지에 있던 product 도메인 관련 클래스를 `services/product/domain` 패키지 하위로 옮김
