+++ app/src/main/java/com/codesoom/assignment/services/product/ProductService.java
@@ -1,8 +1,8 @@
 package com.codesoom.assignment.services.product;
 
 import com.codesoom.assignment.ProductNotFoundException;
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
+import com.codesoom.assignment.services.product.domain.Product;
+import com.codesoom.assignment.services.product.domain.ProductRepository;
 import com.codesoom.assignment.dto.ProductData;
 import org.springframework.stereotype.Service;
