+++ app/src/main/java/com/codesoom/assignment/services/product/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.services.product.domain;
 
 import java.util.List;
 import java.util.Optional;
