+++ app/src/test/java/com/codesoom/assignment/controllers/ProductControllerTest.java
@@ -3,7 +3,7 @@ package com.codesoom.assignment.controllers;
 import com.codesoom.assignment.ProductNotFoundException;
 import com.codesoom.assignment.services.product.ProductService;
 import com.codesoom.assignment.controllers.product.ProductController;
-import com.codesoom.assignment.domain.Product;
+import com.codesoom.assignment.services.product.domain.Product;
 import com.codesoom.assignment.dto.ProductData;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
